Use shared_pointers to store the data
When creating free objects it was assumed that there is an empty
constructor which should not be expected. It is then better to have
a pointer that can be initialised to nullptr and only construct the
object once it is inserted. Though this may have some cost in memory
allocations. To be defined if the compiler optimises it.